refactor: 파이어베이스 관련 함수 분리
